Renames permissions text file
