+++ src/app/modules/item/helpers/permissions-texts.ts
